+++ src/extra/components/pulses.py
@@ -281,9 +281,6 @@ class PulsePattern:
 
         Returns:
             (xarray.DataArray or numpy.ndarray):
-
-        Returns:
-            (numpy.ndarray or pandas.Series):
         """
 
         mask = self._get_pulse_mask()
@@ -1438,10 +1435,43 @@ class PumpProbePulses(XrayPulses, OpticalLaserPulses):
 
         return flags
 
-    @wraps(PulsePattern.pulse_mask)
-    def pulse_mask(self, labelled=True):
-        return TimeserverPulses.pulse_mask(
-            self, labelled=labelled).astype(bool)
+    def pulse_mask(self, labelled=True, field=None):
+        """Get boolean pulse mask.
+
+        The returned mask has the same shape as the full bunch pattern
+        table but only contains boolean flags whether a given pulse
+        was present in this pattern.
+
+        For instance, to see whether each FEL pulse has a corresponding pump
+        laser pulse in a consistent pulse pattern:
+
+        ```python
+        ppl_on = ppp.select_trains(0).pulse_mask(field='ppl').squeeze()
+        fel_on = ppp.select_trains(0).pulse_mask(field='fel').squeeze()
+
+        fel_pulse_is_pumped = ppl_on[fel_on]
+        ```
+
+        Args:
+            labelled (bool, optional): Whether a labelled xarray
+                DataArray (default) or unlabelled numpy array is
+                returned.
+            field (str, optional): 'fel'/'ppl' to mark only FEL or only pump
+                laser pulses in the returned array. By default, pulses are True
+                whether either an FEL pulse or a PPL pulse occured.
+
+        Returns:
+            (xarray.DataArray or numpy.ndarray):
+        """
+        bitfield = TimeserverPulses.pulse_mask(self, labelled=labelled)
+        if field is None:
+            return bitfield.astype(bool)
+        elif field == 'fel':
+            return (bitfield & 1).astype(bool)
+        elif field == 'ppl':
+            return (bitfield & 2).astype(bool)
+        else:
+            raise ValueError(f"{field=!r} parameter was not 'fel'/'ppl'/None")
 
 
 class DldPulses(PulsePattern):